applied clippy flixes

+++ client/src/renderer/mod.rs
@@ -21,7 +21,7 @@ use shader::*;
 use texture::*;
 use batch::*;
 
-use aeonetica_engine::{math::{vector::Vector2, matrix::Matrix4}, collections::OrderedMap, error::{ErrorResult, ErrorValue, IntoError, Fatality, Error}, log};
+use aeonetica_engine::{math::{vector::Vector2, matrix::Matrix4}, collections::OrderedMap, error::{ErrorResult, ErrorValue, IntoError, Fatality, Error}};
 pub(self) use aeonetica_engine::math::camera::Camera;
 
 use self::{sprite_sheet::Sprite, font::BitmapFont, layer::LayerUpdater, pipeline::{Pipeline, DefaultPipeline}, util::Target};
@@ -126,10 +126,8 @@ impl Renderer {
     }
 
     pub(self) fn delete_batch(&mut self, id: &BatchID) {
-        let num_batches = self.batches.len() - 1;
-        self.batches.remove(id).map(|batch| {
-            batch.delete()
-        });
+        let _num_batches = self.batches.len() - 1;
+        if let Some(batch) = self.batches.remove(id) { batch.delete() }
     }
 
     pub fn add_vertices(&mut self, data: &mut VertexData) -> VertexLocation {
